Update diffDrive to O_robot stable version

+++ src/MSP432LISAT/MSP432P401R_LISAT_ccs/Robot/differentialDrive.cpp
@@ -41,7 +41,6 @@ differential_robot_t* robot_init()
     return &robot;
 }
 
-
 float _robot_distance_update_mm(float d_r, float d_l){
 	return (d_r + d_l)/2;
 }
@@ -50,31 +49,32 @@ void _wheel_distance_update_mm( tachometer_t * tachometer){
 	// calculate the distance traveled since the last update
     int32_t ticks = tachometer->ticks;
 	float delta_ticks = ticks - tachometer->prev_ticks;
+
 	// update the previous tachometer ticks
 	tachometer->prev_ticks = ticks;
 	tachometer->delta_dis = (TICK_DIS_NUMERATOR * delta_ticks) / TICK_DIS_DENOMINATOR;  //distance in meter
 }
 
 void robot_position_update(differential_robot_t * robot){
-    _wheel_distance_update_mm( robot->right->tachometer );
-    _wheel_distance_update_mm( robot->left->tachometer );
+    _wheel_distance_update_mm(robot->right->tachometer);
+    _wheel_distance_update_mm(robot->left->tachometer);
 
     float d_r = robot->right->tachometer->delta_dis;
     float d_l = robot->left->tachometer->delta_dis;
-	float d_c = _robot_distance_update_mm(d_r, d_l);
+    float d_c = _robot_distance_update_mm(d_r, d_l);
 
-	float x = robot->pose->x;
-	float y = robot->pose->y;
-	float theta = robot->pose->theta;
+    float x = robot->pose->x;
+    float y = robot->pose->y;
+    float theta = robot->pose->theta;
 
-	x += d_c * cosf(theta);
-	y += d_c * sinf(theta);
-	float delta_theta = (float)(d_r - d_l)/robot->base_len;
-	theta += delta_theta;
+    x += d_c * cos(theta);
+    y += d_c * sin(theta);
+    float delta_theta = (float)(d_r - d_l)/robot->base_len;
+    theta += delta_theta;
 
-	robot->pose->x = x;
-	robot->pose->y = y;
-	robot->pose->theta = atan2f( sinf(theta), cosf(theta));
+    robot->pose->x = x;
+    robot->pose->y = y;
+    robot->pose->theta = atan2(sin(theta), cos(theta));
 
 //	if(cntr_robot_pose_debug >=data_len_robot_pose_debug)
 //	{
@@ -84,7 +84,6 @@ void robot_position_update(differential_robot_t * robot){
 //	robot_y[cntr_robot_pose_debug] = y;
 //	robot_theta[cntr_robot_pose_debug]= robot->pose->theta;
 //	cntr_robot_pose_debug++;
-
 }
